PR-04b: extract broadcasters.rs
Behaviour-preserving relocation:
- broadcasters.rs (new): StatusSnapshot, LayerSource, StatusBroadcaster,
  RestartHandle, PauseBroadcaster, RuntimeEnvironmentBroadcaster,
  ShutdownHandle, wait_for_restart_or_shutdown.

main.rs: 6788 -> 6588 LOC.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
