+++ src/daemon/main.rs
@@ -95,6 +95,7 @@ mod config;
 mod focus;
 mod args;
 mod autostart;
+mod broadcasters;
 
 use constants::*;
 use errors::DynError;
@@ -104,10 +105,11 @@ use config::*;
 use focus::*;
 use args::*;
 use autostart::*;
+use broadcasters::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*};
 
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -265,213 +267,6 @@ async fn send_control_command_broadcast(
     });
     let results = futures_util::future::join_all(futures).await;
     Ok(BroadcastReport { results })
-}
-
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct StatusSnapshot {
-    layer: String,
-    virtual_keys: Vec<String>,
-    layer_source: LayerSource,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum LayerSource {
-    Focus,
-    External,
-}
-
-impl LayerSource {
-    fn as_str(&self) -> &'static str {
-        match self {
-            LayerSource::Focus => "focus",
-            LayerSource::External => "external",
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct StatusBroadcaster {
-    sender: watch::Sender<StatusSnapshot>,
-}
-
-#[derive(Clone, Debug)]
-struct RestartHandle {
-    sender: watch::Sender<bool>,
-}
-
-#[derive(Clone, Debug)]
-struct PauseBroadcaster {
-    sender: watch::Sender<bool>,
-}
-
-#[derive(Clone, Debug)]
-struct RuntimeEnvironmentBroadcaster {
-    sender: watch::Sender<Environment>,
-}
-
-#[derive(Clone, Debug)]
-struct ShutdownHandle {
-    sender: watch::Sender<bool>,
-}
-
-impl RestartHandle {
-    fn new() -> Self {
-        let (sender, _) = watch::channel(false);
-        Self { sender }
-    }
-
-    fn subscribe(&self) -> watch::Receiver<bool> {
-        self.sender.subscribe()
-    }
-
-    fn request(&self) {
-        self.sender.send_replace(true);
-    }
-}
-
-impl ShutdownHandle {
-    fn new() -> Self {
-        let (sender, _) = watch::channel(false);
-        Self { sender }
-    }
-
-    fn subscribe(&self) -> watch::Receiver<bool> {
-        self.sender.subscribe()
-    }
-
-    fn request(&self) {
-        self.sender.send_replace(true);
-    }
-}
-
-impl PauseBroadcaster {
-    fn new() -> Self {
-        let (sender, _) = watch::channel(false);
-        Self { sender }
-    }
-
-    fn subscribe(&self) -> watch::Receiver<bool> {
-        self.sender.subscribe()
-    }
-
-    fn is_paused(&self) -> bool {
-        *self.sender.borrow()
-    }
-
-    fn set_paused(&self, paused: bool) -> bool {
-        let current = *self.sender.borrow();
-        if current == paused {
-            return false;
-        }
-        self.sender.send_replace(paused);
-        true
-    }
-}
-
-impl RuntimeEnvironmentBroadcaster {
-    fn new(initial: Environment) -> Self {
-        let (sender, _) = watch::channel(initial);
-        Self { sender }
-    }
-
-    fn current(&self) -> Environment {
-        *self.sender.borrow()
-    }
-
-    fn set_current(&self, env: Environment) {
-        self.sender.send_replace(env);
-    }
-
-    fn subscribe(&self) -> watch::Receiver<Environment> {
-        self.sender.subscribe()
-    }
-}
-
-impl StatusBroadcaster {
-    fn new() -> Self {
-        let initial = StatusSnapshot {
-            layer: String::new(),
-            virtual_keys: Vec::new(),
-            layer_source: LayerSource::External,
-        };
-        let (sender, _) = watch::channel(initial);
-        Self { sender }
-    }
-
-    fn subscribe(&self) -> watch::Receiver<StatusSnapshot> {
-        self.sender.subscribe()
-    }
-
-    fn snapshot(&self) -> StatusSnapshot {
-        self.sender.borrow().clone()
-    }
-
-    fn update_layer(&self, layer: String, source: LayerSource) {
-        self.update(|state| {
-            state.layer = layer;
-            state.layer_source = source;
-        });
-    }
-
-    fn update_virtual_keys(&self, virtual_keys: Vec<String>) {
-        self.update(|state| {
-            state.virtual_keys = virtual_keys;
-        });
-    }
-
-    fn update_focus_layer(&self, layer: String) {
-        let mut next = self.sender.borrow().clone();
-        next.layer = layer;
-        next.layer_source = LayerSource::Focus;
-        self.sender.send_replace(next);
-    }
-
-    fn set_paused_status(&self, layer: String) {
-        let mut next = self.sender.borrow().clone();
-        next.layer = layer;
-        next.layer_source = LayerSource::External;
-        next.virtual_keys = Vec::new();
-        self.sender.send_replace(next);
-    }
-
-    fn update<F>(&self, updater: F)
-    where
-        F: FnOnce(&mut StatusSnapshot),
-    {
-        let current = self.sender.borrow().clone();
-        let mut next = current.clone();
-        updater(&mut next);
-        if next != current {
-            self.sender.send_replace(next);
-        }
-    }
-}
-
-async fn wait_for_restart_or_shutdown(
-    restart_handle: &RestartHandle,
-    shutdown_handle: &ShutdownHandle,
-) -> RunOutcome {
-    let mut restart_receiver = restart_handle.subscribe();
-    let mut shutdown_receiver = shutdown_handle.subscribe();
-
-    if *shutdown_receiver.borrow() {
-        return RunOutcome::Exit;
-    }
-    if *restart_receiver.borrow() {
-        return RunOutcome::Restart;
-    }
-
-    tokio::select! {
-        _ = shutdown_receiver.changed() => RunOutcome::Exit,
-        _ = restart_receiver.changed() => {
-            if *shutdown_receiver.borrow() {
-                RunOutcome::Exit
-            } else {
-                RunOutcome::Restart
-            }
-        }
-    }
 }
 
 // === SNI Indicator ===
